+++ src/console/controllers/BlockTypeGroupsController.php
@@ -3,8 +3,10 @@
 namespace benf\neo\console\controllers;
 
 use benf\neo\enums\BlockTypeGroupDropdown;
+use benf\neo\Plugin as Neo;
 use Craft;
 use craft\console\Controller;
+use craft\db\Query;
 use craft\helpers\Console;
 use craft\helpers\Db;
 use yii\console\ExitCode;
@@ -22,6 +24,11 @@ class BlockTypeGroupsController extends Controller
      * @var int|null A block type group ID.
      */
     public ?int $groupId = null;
+
+    /**
+     * @var bool Whether to delete block types belonging to the block type group.
+     */
+    public bool $deleteBlockTypes = false;
 
     /**
      * @var ?int A new name to set for the block type group.
@@ -45,7 +52,10 @@ class BlockTypeGroupsController extends Controller
     {
         $options = parent::options($actionID);
 
-        if ($actionID === 'edit') {
+        if ($actionID === 'delete') {
+            $options[] = 'groupId';
+            $options[] = 'deleteBlockTypes';
+        } elseif ($actionID === 'edit') {
             $options[] = 'groupId';
             $options[] = 'setName';
             $options[] = 'blankName';
@@ -53,6 +63,54 @@ class BlockTypeGroupsController extends Controller
         }
 
         return $options;
+    }
+
+    /**
+     * Deletes a Neo block type group.
+     *
+     * @return int
+     */
+    public function actionDelete(): int
+    {
+        if (!$this->groupId) {
+            $this->stderr('The --group-id option must be specified.' . PHP_EOL, Console::FG_RED);
+            return ExitCode::USAGE;
+        }
+
+        $projectConfig = Craft::$app->getProjectConfig();
+        $group = Neo::$plugin->blockTypes->getGroupById($this->groupId);
+        $blockTypeUids = (new Query())
+            ->select(['uid'])
+            ->from('{{%neoblocktypes}}')
+            ->where(['groupId' => $this->groupId])
+            ->column();
+
+        if ($this->deleteBlockTypes) {
+            $this->stdout('Deleting the group\'s block types...' . PHP_EOL);
+            foreach ($blockTypeUids as $blockTypeUid) {
+                $projectConfig->remove('neoBlockTypes.' . $blockTypeUid);
+            }
+        } elseif (!empty($blockTypeUids)) {
+            $this->stdout('Reassigning the group\'s block types to the field\'s previous group...' . PHP_EOL);
+            $prevGroupUid = (new Query())
+                ->select(['uid'])
+                ->from(['{{%neoblocktypegroups}}'])
+                ->where(['fieldId' => $group->fieldId])
+                ->andWhere(['<', 'sortOrder', $group->sortOrder])
+                ->orderBy(['sortOrder' => SORT_DESC])
+                ->scalar();
+
+            foreach ($blockTypeUids as $blockTypeUid) {
+                $projectConfig->set('neoBlockTypes.' . $blockTypeUid . '.group', $prevGroupUid);
+            }
+        }
+
+        // Now we can delete the group
+        $this->stdout('Deleting the group...' . PHP_EOL);
+        $projectConfig->remove('neoBlockTypeGroups.' . $group->uid);
+        $this->stdout('Done.' . PHP_EOL);
+
+        return ExitCode::OK;
     }
 
     /**