+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- Added the `php craft neo/block-type-groups/delete` console command
 - Added the `php craft neo/block-type-groups/edit` console command
 - Added `benf\neo\console\controllers\BlockTypeGroupsController`
 - Added `benf\neo\enums\BlockTypeGroupDropdown`